Refactor the back-end utility directory to group related files better
