+++ apps/back-end/src/server/app.ts
@@ -4,7 +4,7 @@ import express from "express";
 import { resolveErrors } from "src/server/errors";
 import createEndpoints from "src/server/routes";
 import ALLOWED_ORIGINS from "src/utility/constants/ALLOWED_ORIGINS";
-import setupCors from "src/utility/setupCors";
+import setupCors from "src/utility/initialisers/setupCors";
 import "src/instrument";
 
 const app = express();
