+++ apps/back-end/src/server/errors/handleUnfoundEndpoint.ts
@@ -1,5 +1,5 @@
-import endpointNotFoundError from "src/utility/endpointNotFoundError";
-import handleFallthroughMiddleware from "src/utility/handleFallthroughMiddleware";
+import endpointNotFoundError from "src/utility/errors/endpointNotFoundError";
+import handleFallthroughMiddleware from "src/utility/handlers/handleFallthroughMiddleware";
 
 const handleUnfoundEndpoint = handleFallthroughMiddleware(async (request) => {
   // If we ever get into this app.use, the endpoint does not exist
