+++ apps/back-end/src/server/routes/auth/index.ts
@@ -4,7 +4,7 @@ import getAuthGoogle from "src/server/routes/auth/getAuthGoogle";
 import getAuthGoogleCallback from "src/server/routes/auth/getAuthGoogleCallback";
 import postAuthEndToEnd from "src/server/routes/auth/postAuthEndToEnd";
 import postAuthLogout from "src/server/routes/auth/postAuthLogout";
-import registerEndpoints from "src/utility/registerEndpoints";
+import registerEndpoints from "src/utility/initialisers/registerEndpoints";
 
 function initialiseAuthRouter(router: Router) {
   registerEndpoints(router, {
