+++ apps/back-end/src/server/routes/blogs/index.ts
@@ -4,7 +4,7 @@ import getBlogById from "src/server/routes/blogs/getBlogById";
 import getBlogs from "src/server/routes/blogs/getBlogs";
 import postBlogs from "src/server/routes/blogs/postBlogs";
 import putBlogById from "src/server/routes/blogs/putBlogById";
-import registerEndpoints from "src/utility/registerEndpoints";
+import registerEndpoints from "src/utility/initialisers/registerEndpoints";
 
 function initialiseBlogsRouter(router: Router) {
   registerEndpoints(router, {
