+++ apps/back-end/src/server/routes/blogs/putBlogById.ts
@@ -6,10 +6,10 @@ import { parseEditBlogData } from "@lexicon/models";
 
 import { getConnection } from "src/database/connection";
 import { changeBlogState, editBlog, selectBlogView } from "src/services/blogs";
-import handleEndpointMiddleware from "src/utility/handleEndpointMiddleware";
-import resourceNotFoundError from "src/utility/resourceNotFoundError";
-import requireAuth from "src/utility/validators/requireAuth";
-import validateUUID from "src/utility/validators/validateUUID";
+import resourceNotFoundError from "src/utility/errors/resourceNotFoundError";
+import handleEndpointMiddleware from "src/utility/handlers/handleEndpointMiddleware";
+import requireAuth from "src/utility/handlers/requireAuth";
+import validateUUID from "src/utility/handlers/validateUUID";
 
 function putBlogById(blogs: Router) {
   blogs.param("blogId", validateUUID).put(
