+++ apps/back-end/src/server/routes/currentUser/index.ts
@@ -2,7 +2,7 @@ import type { Router } from "express";
 
 import getCurrentUser from "src/server/routes/currentUser/getCurrentUser";
 import putCurrentUserProfile from "src/server/routes/currentUser/putCurrentUserProfile";
-import registerEndpoints from "src/utility/registerEndpoints";
+import registerEndpoints from "src/utility/initialisers/registerEndpoints";
 
 function initialiseCurrentUserRouter(router: Router) {
   registerEndpoints(router, {
