+++ apps/back-end/src/server/routes/users/index.ts
@@ -1,7 +1,7 @@
 import type { Router } from "express";
 
 import getUserById from "src/server/routes/users/getUserById";
-import registerEndpoints from "src/utility/registerEndpoints";
+import registerEndpoints from "src/utility/initialisers/registerEndpoints";
 
 function initialiseUsersRouter(router: Router) {
   registerEndpoints(router, {
