+++ apps/back-end/src/utility/constants/ENV.ts
@@ -1,4 +1,4 @@
-import parseEnv from "src/utility/parseEnv";
+import parseEnv from "src/utility/miscellaneous/parseEnv";
 
 const ENV = parseEnv(process.env.NODE_ENV ?? "development");
