+++ apps/back-end/src/utility/errors/authRequiredError.ts
