+++ apps/back-end/src/utility/errors/endpointNotFoundError.ts
