+++ apps/back-end/src/utility/errors/resourceNotFoundError.ts
