+++ apps/back-end/src/utility/handlers/validateUUID.ts
